Cognito's token verification function added
1. Cognito's token verification function added
2. Fixed bug when changing profile

Note:
Add .env value
UPLOAD_OPERATOR_PROFILE_DIR
UPLOAD_SUBSCRIPTION_DIR
MODE_DEV
AWS_COGNITO_USERPOOL_ID
AWS_COGNITO_REGION
AWS_COGNITO_USERPOOL_WEBCLIENT_ID

+++ src/graphql/resolver/operatorResolver.ts
@@ -1,25 +1,30 @@
 import { OperatorModel } from "../../models/dbmodel";
+const global = require('../../global');
 
 import fs from 'fs';
 import path from 'path';
 
 const operatorResolver = {
     Query: {
-      async operatorCount (_parent: any, _args: any) {
+      async operatorCount (_parent: any, _args: any, { headers }: any) {
+        await global.isAuthorization(headers);
         const count = await OperatorModel.countDocuments(_args);
         return count;
       },
-      async operators (_parent: any, _args: any) {
+      async operators (_parent: any, _args: any, { headers }: any) {
+        await global.isAuthorization(headers);
         const lists = await OperatorModel.find();
         return lists;
       },
-      async operator (_parent: any, _args: any) {
+      async operator (_parent: any, _args: any, { headers }: any) {
+        await global.isAuthorization(headers);
         const result = await OperatorModel.findOne({username: _args.username});
         return result;
       }
     },
     Mutation: {
-        async postOperator(_parent: any, _args: any) {
+        async postOperator(_parent: any, _args: any, { headers }: any) {
+          await global.isAuthorization(headers);
           const newData = {
             ..._args.input,
           }
@@ -33,28 +38,39 @@ const operatorResolver = {
           });
           return result;
         },
-        async updateOperator(_parent: any, _args: any) {
+        async updateOperator(_parent: any, _args: any, { headers }: any) {
+          await global.isAuthorization(headers);
           const updateData = {
             ..._args.input,
           }
           let newOperator
-          let _path = path.join(path.resolve(), updateData.profileImage)
-          if (fs.existsSync(_path)) {
-            var fileName = path.basename(_path)
-            let fileType = path.extname(fileName)
-            const newFileName = _args.username + Date.now() + fileType
-            var newPath = path.join(path.resolve(), `${process.env.UPLOAD_PROFILE_DIR}${newFileName}`)
+          if (updateData.profileImage){
+            let _path = path.join(path.resolve(), updateData.profileImage)
+            if (fs.existsSync(_path)) {
+              var fileName = path.basename(_path)
+              let fileType = path.extname(fileName)
+              const newFileName = _args.username + Date.now() + fileType
+              const operatorDirPath = `${process.env.UPLOAD_OPERATOR_PROFILE_DIR}`
+              fs.mkdirSync(path.join(path.resolve(), operatorDirPath), { recursive: true });
+              var newPath = path.join(path.resolve(), `${operatorDirPath}${newFileName}`)
+              newOperator = {
+                ...updateData,
+                profileImage:`${operatorDirPath}${newFileName}`,
+              }
+              fs.rename(_path, newPath, function (err) {
+                if (err) throw err
+                // delete temp directory
+                fs.rm(path.dirname(_path), { recursive: true }, (err) => {});              
+              })
+            }else{
+              newOperator = {
+                ...updateData,
+              }
+            }
+          }else{
             newOperator = {
               ...updateData,
-              profileImage:`${process.env.UPLOAD_PROFILE_DIR}${newFileName}`,
             }
-            fs.rename(_path, newPath, function (err) {
-              if (err) throw err
-              // delete temp directory
-              fs.rm(path.dirname(_path), { recursive: true }, (err) => {});              
-            })
-          }else{
-            
           }
           if (await OperatorModel.findOne({username: _args.username})) {
             let results;
@@ -77,7 +93,8 @@ const operatorResolver = {
             return results;
           }
         },
-        async deleteOperator(_parent: any, _args: any) {
+        async deleteOperator(_parent: any, _args: any, { headers }: any) {
+          await global.isAuthorization(headers);
           let results;
           await OperatorModel.findOneAndDelete({username:_args.username})
           .then((result) => {
